third_party/WebKit/LayoutTests/fullscreen/rendering/overflow.html

<!DOCTYPE html>
<title>overflow of root element when descendant is fullscreen</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<style>:root { overflow: scroll; }</style>
<div id="log"></div>
<script>
async_test(function(t)
{
    var root = document.documentElement;
    trusted_request(t, document.querySelector("div"));

    document.addEventListener("fullscreenchange", t.step_func_done(function()
    {
        // The scrollbar should be removed.
        assert_equals(getComputedStyle(root).overflow, "hidden");
        assert_equals(root.clientWidth, window.innerWidth);
    }));
});
</script>
